Move App to containers dir
